manager/pyproject.toml
[tool.poetry]
name = "manager"
version = "0.1.0"
description = ""
authors = ["corgi-in-tights <59304120+corgi-in-tights@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pvporcupine = "^3.0.5"
websockets = "^15.0.1"
pyaudio = "^0.2.14"
deepgram-sdk = "^4.6.0"
sounddevice = "^0.5.2"
numpy = "^2.3.1"
python-dotenv = "^1.1.1"
scipy = "^1.16.0"
openai = "^1.97.0"
pyserial = "^3.5"
motor = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
